delay v0.04
Working version of delay function. Processes full matrix and retruns 4th ordered sinc filter with cos^2 weighting filter.

The returned matrix does not contain the original signal. I assumed that was how the function worked. I will test with the Array Toolbox and fix if that is not the case.
 Changes to be committed:
	modified:   src/delay.cpp
	modified:   src/driver
	modified:   src/main.cpp